Rename instrumentors pkg to instrumentation
