+++ internal/pkg/instrumentation/bpf/google.golang.org/grpc/bpf_bpfel_x86.go
