+++ internal/pkg/instrumentation/bpf/google.golang.org/grpc/probe.go
