+++ internal/pkg/instrumentation/utils/kernel.go
